Getting the build to work on Mono 2.10 and ignoring files produced by MonoDevelop
Getting rid of the ProcessAffinity helper class.  According to this
MSDN blog post --
http://blogs.microsoft.co.il/blogs/sasha/archive/2008/04/20/parallelism-
and-cpu-affinity.aspx -- this is the wrong way to set thread affinity
in .NET.  If in the future we need to be able to do this we should just
use the Process.ProcessorAffinity property.

But now all the project build and the unit tests pass on Mono.