Use a pointer in MultiProducerSequencer
The goal is to improve the sequencer performance by removing the bound
checks. The Java version used UNSAFE here for this purpose.

It was previously possible to pin the buffer using GCHandle or
AllocHGlobal, but then the MultiProducerSequencer would have to
implement IDisposable to cleanup the handle / memory.

The POH is a good way to get a safe pointer without changing anything to
the sequencer API.