Support timeouts in AsyncWaitStrategy
AsyncWaitStrategy was built using a nested IWaitStrategy which was
used for the synchronous waits. It was slightly inefficient because
two locks were taken when the nested strategy was a blocking strategy.
Also, it did not support timeouts in the async path.

The new implementation is based on TimeoutBlockingWaitStrategy
and supports timeouts in the async path.