Zero-extend int
Following Mendel's link to "Peanut butter" https://devblogs.microsoft.com/dotnet/performance-improvements-in-net-6/#peanut-butter.

This gives 2-3% on my machine on the same HT core. The bench with different physical cores became very unstable on my machine: jumping from 1xx to 280 (compared to my previous PR, which works as it did on the same machine).

Adding `ExtraWork` method, which was removed for some reason, every time is tedious so I have not, but it makes batching behavior more stable. So these numbers need to be double checked on different machines.