Replace GetBarrierFor by GetDependentSequencesFor
Sequence barriers are stateful components which are dangerous to expose,
but they were the only option to identify the source processor in a wait
strategy and track the dependent sequences of a given processor.

Exposing the dependent sequences instead of the barrier seems to
address both issues (the barriers were already replaced by dependent
sequences in the wait strategy API in previous commits).